Move remaining zoekt code out of internal/search/run
